+++ ChangeLog.md
@@ -10,6 +10,7 @@
 	- ttp.pl alert: fix cmd vs. sh execution and quotings
     - logicals regular expression is no more an array as it can satisfies itself with a single string (todo #87)
     - Change TTP::commandExec() prototype to more standard '$command, $opts' (todo #86)
+    - Deduplicate TTP::runnerDaemon::_mqtt_timeout() and TTP::runnerDaemon::messagingTimeout() (todo #83)
 
 ### 4.4.1