+++ ChangeLog.md
@@ -14,6 +14,7 @@
     - Bump minor candidate version number due to site.json change
     - RunnerDaemon better check listeningPort (todo #81) and listeningInterval (todo #82)
     - sufix is renamed suffix (todo #80)
+    - TTP::commandByOs() now accepts 'withCommands' option
 
 ### 4.4.1